+++ cmd/hash.go
@@ -8,56 +8,110 @@ import (
 	"fmt"
 	"strings"
 
+	"github.com/matthewhartstonge/argon2"
 	log "github.com/sirupsen/logrus"
 	"github.com/spf13/cobra"
 	"github.com/tommi2day/gomodules/pwlib"
 	"golang.org/x/crypto/bcrypt"
 )
 
-var hashCmd = &cobra.Command{
-	Use:   "hash",
-	Short: "command to hashing Passwords ",
-	Long: `prepare a password hash
-currently supports basic auth(for http), md5 and scram(for postgresql),SSHA(for LDAP) and bcrypt(for htpasswd) methods`,
-	RunE:         doHash,
-	SilenceUsage: true,
-}
-
 const mMD5 = "md5"
 const mScram = "scram"
 const mBcrypt = "bcrypt"
 const mSSHA = "ssha"
 const mBasic = "basic"
+const mArgon2 = "argon2"
 
-func init() {
-	hashCmd.Flags().StringP("username", "u", "", "username for basic, scram and md5 hash")
-	hashCmd.Flags().StringP("password", "p", "", "password to encrypt")
-	hashCmd.Flags().StringP("prefix", "P", "", "prefix for hash string(default basic='Authorization: Basic ',md5={MD5},ssha={SSHA})")
-	hashCmd.Flags().StringP("test", "T", "", "test given hash to verify against hashed password (not for scram)")
-	hashCmd.Flags().StringP("hash-method", "M", "", "method to use for hashing, one of basic, md5, scram, bcrypt, ssha")
-	_ = hashCmd.MarkFlagRequired("password")
-	_ = hashCmd.MarkFlagRequired("hash-method")
-	hashCmd.SetHelpFunc(hideFlags)
-	RootCmd.AddCommand(hashCmd)
-	// hide unused flags, do not on group command
+var hashCmd = &cobra.Command{
+	Use:   "hash",
+	Short: "command to hashing Passwords ",
+	Long: `prepare a password hash
+currently supports basic auth(for http), md5 and scram(for postgresql),SSHA(for LDAP), bcrypt(for htpasswd) and argon2(for vaultwarden)`,
 }
 
-func doHash(cmd *cobra.Command, _ []string) error {
-	m, _ := cmd.Flags().GetString("hash-method")
-	switch m {
-	case mMD5:
-		return hashMD5(cmd, nil)
-	case mScram:
-		return hashScram(cmd, nil)
-	case mBcrypt:
-		return hashBcrypt(cmd, nil)
-	case mSSHA:
-		return hashSSHA(cmd, nil)
-	case mBasic:
-		return basicAuth(cmd, nil)
-	default:
-		return fmt.Errorf("method %s not supported", m)
-	}
+var md5Cmd = &cobra.Command{
+	Use:          mMD5,
+	Short:        "command to hashing User/Password with MD5 method",
+	RunE:         hashMD5,
+	SilenceUsage: true,
+}
+var scramCmd = &cobra.Command{
+	Use:          mScram,
+	Short:        "command to hashing User/Password with SCRAM method",
+	RunE:         hashScram,
+	SilenceUsage: true,
+}
+
+var bcryptCmd = &cobra.Command{
+	Use:          mBcrypt,
+	Short:        "command to hashing Passwords with BCrypt method",
+	RunE:         hashBcrypt,
+	SilenceUsage: true,
+}
+
+var sshaCmd = &cobra.Command{
+	Use:          mSSHA,
+	Short:        "command to hashing Passwords with SSHA method",
+	RunE:         hashSSHA,
+	SilenceUsage: true,
+}
+
+var basicCmd = &cobra.Command{
+	Use:          mBasic,
+	Short:        "command to encoding User/Password with HTTP Basic method",
+	RunE:         basicAuth,
+	SilenceUsage: true,
+}
+
+var argon2Cmd = &cobra.Command{
+	Use:          mArgon2,
+	Short:        "command to hashing Passwords with Argon2 method",
+	RunE:         hashArgon2,
+	SilenceUsage: true,
+}
+
+func init() {
+	// hide unused flags, do not on group command
+	hashCmd.SetHelpFunc(hideFlags)
+	RootCmd.AddCommand(hashCmd)
+
+	md5Cmd.Flags().StringP("username", "u", "", "username ")
+	md5Cmd.Flags().StringP("password", "p", "", "password to encrypt")
+	md5Cmd.Flags().StringP("prefix", "P", "", "prefix for string(default md5={MD5}")
+	md5Cmd.Flags().StringP("test", "T", "", "test given hash to verify against hashed password")
+	_ = md5Cmd.MarkFlagRequired("password")
+	_ = md5Cmd.MarkFlagRequired("username")
+	hashCmd.AddCommand(md5Cmd)
+
+	scramCmd.Flags().StringP("username", "u", "", "username")
+	scramCmd.Flags().StringP("password", "p", "", "password to encrypt")
+	scramCmd.Flags().StringP("prefix", "P", "", "prefix for hash string(default basic='Authorization: Basic ',md5={MD5},ssha={SSHA})")
+	_ = scramCmd.MarkFlagRequired("password")
+	_ = scramCmd.MarkFlagRequired("username")
+	hashCmd.AddCommand(scramCmd)
+
+	basicCmd.Flags().StringP("username", "u", "", "username")
+	basicCmd.Flags().StringP("password", "p", "", "password to encode")
+	basicCmd.Flags().StringP("prefix", "P", "", "prefix for hash string(default basic='Authorization: Basic ')")
+	basicCmd.Flags().StringP("test", "T", "", "test given hash to verify against encoded password")
+	_ = basicCmd.MarkFlagRequired("password")
+	_ = basicCmd.MarkFlagRequired("username")
+	hashCmd.AddCommand(basicCmd)
+
+	bcryptCmd.Flags().StringP("password", "p", "", "password to encode")
+	bcryptCmd.Flags().StringP("test", "T", "", "test given hash to verify against encoded password")
+	_ = bcryptCmd.MarkFlagRequired("password")
+	hashCmd.AddCommand(bcryptCmd)
+
+	sshaCmd.Flags().StringP("password", "p", "", "password to encode")
+	sshaCmd.Flags().StringP("test", "T", "", "test given hash to verify against encoded password")
+	_ = md5Cmd.MarkFlagRequired("password")
+	hashCmd.AddCommand(sshaCmd)
+
+	argon2Cmd.Flags().StringP("password", "p", "", "password to encode")
+	argon2Cmd.Flags().StringP("test", "T", "", "test given hash to verify against encoded password")
+	_ = argon2Cmd.MarkFlagRequired("password")
+	hashCmd.AddCommand(argon2Cmd)
 }
 
 func basicAuth(cmd *cobra.Command, _ []string) error {
@@ -232,5 +286,49 @@ func doSSHA(sshaPlain string, prefix string) (result string, err error) {
 	enc := pwlib.SSHAEncoder{}
 	encoded, err := enc.Encode([]byte(sshaPlain), prefix)
 	result = string(encoded)
+	return
+}
+
+func hashArgon2(cmd *cobra.Command, _ []string) error {
+	var err error
+	var argon2Value string
+	password, _ := cmd.Flags().GetString("password")
+	if password == "" {
+		err = fmt.Errorf("password is required")
+		return err
+	}
+	test, _ := cmd.Flags().GetString("test")
+	// generate only if test is empty
+	if test == "" {
+		argon2Value, err = doArgon2(password)
+		if err != nil {
+			return fmt.Errorf("error while hashing password:%s", err)
+		}
+		log.Infof("ARGON2 hash for password '%s' is '%s'", password, argon2Value)
+		cmd.Println(argon2Value)
+		return nil
+	}
+	// compare only if test is not empty
+	ok, err := argon2.VerifyEncoded([]byte(password), []byte(test))
+	if err != nil {
+		return fmt.Errorf("argon2 compare failed:%s", err)
+	}
+
+	if !ok {
+		log.Infof("ERROR, test input does not match argon2 hash '%s'", argon2Value)
+		return fmt.Errorf("ERROR, test input does not match ssha hash '%s'", argon2Value)
+	}
+	log.Infof("OK, test input matches argon2 hash")
+	cmd.Println("OK, test input matches argon2 hash")
+	return nil
+}
+
+func doArgon2(password string) (result string, err error) {
+	argon := argon2.DefaultConfig()
+	// Waaahht??! It includes magic salt generation for me ! Yasss...
+	encoded, err := argon.HashEncoded([]byte(password))
+	if err == nil {
+		result = string(encoded)
+	}
 	return
 }