+++ cmd/hash_test.go
@@ -19,6 +19,7 @@ const testBcrypt = "$2a$10$Y3xlpzHMnNyZXm.rnIGqouf9NpPP.OCtB6FakJC3nK/Z1CYmC3Amq
 const testMD5 = "{MD5}ebcd5bc0483385f278b814600272d794"
 const testSSHA = "{SSHA}r3myNFUMmkpxkaJ9EIr071i9x+1MqPgS"
 const testBasic = "dGVzdEhhc2hVc2VybmFtZTp0ZXN0SGFzaFBhc3N3b3Jk"
+const testArgon2 = "$argon2id$v=19$m=65536,t=3,p=4$yVSLalsV0ZyoyByE5IQDVg$V14dRnxoKArosnameO3QdnFstLMbGvqHhJsUbZ9UQcI"
 
 func TestHash(t *testing.T) {
 	var out string
@@ -34,7 +35,7 @@ func TestHash(t *testing.T) {
 	t.Run("TestHashSSHA", func(t *testing.T) {
 		args := []string{
 			"hash",
-			"--hash-method=ssha",
+			"ssha",
 			"--password", hashPassword,
 			"--info",
 			"--unit-test",
@@ -47,7 +48,7 @@ func TestHash(t *testing.T) {
 	t.Run("TestHashSSHAMatch", func(t *testing.T) {
 		args := []string{
 			"hash",
-			"--hash-method=ssha",
+			"ssha",
 			"--password", hashPassword,
 			"--test", testSSHA,
 			"--info",
@@ -58,10 +59,11 @@ func TestHash(t *testing.T) {
 		assert.Contains(t, out, "OK, test input matches", "Output should contain OK message")
 		t.Logf(out)
 	})
+	_ = md5Cmd.Flags().Set("test", "")
 	t.Run("TestHashMD5", func(t *testing.T) {
 		args := []string{
 			"hash",
-			"--hash-method=md5",
+			"md5",
 			"--username", hashUsername,
 			"--password", hashPassword,
 			"--prefix", "md5",
@@ -77,7 +79,7 @@ func TestHash(t *testing.T) {
 	t.Run("TestHashMD5Match", func(t *testing.T) {
 		args := []string{
 			"hash",
-			"--hash-method=md5",
+			"md5",
 			"--username", hashUsername,
 			"--password", hashPassword,
 			"--prefix", "md5",
@@ -93,7 +95,7 @@ func TestHash(t *testing.T) {
 	t.Run("TestHashScram", func(t *testing.T) {
 		args := []string{
 			"hash",
-			"--hash-method=scram",
+			"scram",
 			"--username", hashUsername,
 			"--password", hashPassword,
 			"--info",
@@ -115,7 +117,7 @@ func TestHash(t *testing.T) {
 	t.Run("TestHashBcrypt", func(t *testing.T) {
 		args := []string{
 			"hash",
-			"--hash-method=bcrypt",
+			"bcrypt",
 			"--password", hashPassword,
 			"--info",
 			"--unit-test",
@@ -129,7 +131,7 @@ func TestHash(t *testing.T) {
 	t.Run("TestHashBcryptMatch", func(t *testing.T) {
 		args := []string{
 			"hash",
-			"--hash-method=bcrypt",
+			"bcrypt",
 			"--password", hashPassword,
 			"--test", testBcrypt,
 			"--info",
@@ -141,11 +143,10 @@ func TestHash(t *testing.T) {
 		t.Logf(out)
 	})
 
-	_ = hashCmd.Flags().Set("test", "")
 	t.Run("TestHashBasic", func(t *testing.T) {
 		args := []string{
 			"hash",
-			"--hash-method=basic",
+			"basic",
 			"--username", hashUsername,
 			"--password", hashPassword,
 			"--prefix", "Authorization: Basic ",
@@ -161,7 +162,7 @@ func TestHash(t *testing.T) {
 	t.Run("TestHashBasicMatch", func(t *testing.T) {
 		args := []string{
 			"hash",
-			"--hash-method=basic",
+			"basic",
 			"--username", hashUsername,
 			"--password", hashPassword,
 			"--test", testBasic,
@@ -170,6 +171,36 @@ func TestHash(t *testing.T) {
 		}
 		out, err = common.CmdRun(RootCmd, args)
 		require.NoErrorf(t, err, "hash basic command should  not return an error:%s", err)
+		assert.Contains(t, out, "OK, test input matches", "Output should contain OK message")
+		t.Logf(out)
+	})
+
+	_ = hashCmd.Flags().Set("test", "")
+	t.Run("TestHashArgon2", func(t *testing.T) {
+		args := []string{
+			"hash",
+			"argon2",
+			"--password", hashPassword,
+			"--info",
+			"--unit-test",
+		}
+		out, err = common.CmdRun(RootCmd, args)
+		require.NoErrorf(t, err, "hash argon2 command should  not return an error:%s", err)
+		assert.Contains(t, out, "$argon2id$", "Output should contain argon header")
+		t.Logf(out)
+	})
+
+	t.Run("TestHashArgon2tMatch", func(t *testing.T) {
+		args := []string{
+			"hash",
+			"argon2",
+			"--password", hashPassword,
+			"--test", testArgon2,
+			"--info",
+			"--unit-test",
+		}
+		out, err = common.CmdRun(RootCmd, args)
+		require.NoErrorf(t, err, "hash argon2 command should  not return an error:%s", err)
 		assert.Contains(t, out, "OK, test input matches", "Output should contain OK message")
 		t.Logf(out)
 	})